Update patch set 6
Patch Set 6:

Dean Troyer Thanks for your comment.

I suggest to implement it not as an extras.d plugin only because HBase needs for Ceilometer as backend.
This sequence is critical for Ceilometer. When Ceilometer starts it expects that hbase does work and available. I mean that part of code
https://github.com/openstack/ceilometer/blob/master/ceilometer/storage/impl_hbase.py

Do you think it is reasonable to keep current implementation taking into account Ceilometer's needs?
Thanks.

Patch-set: 6
Label: Code-Review=0