extra.d to use Sheepdog as devstack backing store.
To use, add "enable_service sheepdog" to your config.

This configures a test Sheepdog cluster with all sheepdog processes
running on the same machine. This uses the "-c local" option for sheep,
which runs the cluster-membership protocol over unix sockets. This is
not intended for production use. Glance and Cinder are then configured
to use the Sheepdog cluster.

For all configurable options, see lib/storage/sheepdog.

Change-Id: I05563b049c33fce3939eb90b6deaee41a78a3532