Update patch set 1
Patch Set 1: I would prefer that you didn't merge this

(1 inline comment)

So 'key' was used rather than keystone because 'key' is the name of the service in ENABLED_SERVICES and it seemed better to keep it consistent. 

I don't mind either way, but given that anyone who cares about this has already either figured it out or moved on might it be easier to change the docs?

Patch-set: 1
Label: Code-Review=-1