Only enable heat for libvirt
Glance will reject images that it doesn't understand, and only libvirt can
use qcow2 images.  This means for systems without qcow2 support, devstack
will fail to run until this change is merged.
Further, heat requires this qcow2 image to be present, to heat should not
be enabled on platforms which do not support qcow2.

Change-Id: Ic92815f3d32880e987456c0428953dddaa81c94c