Update patch set 2
Patch Set 2:

I think we still need the fix.

For single string, we do not need to set values in devstack and can use default values.

But for multiple string option, it is better to set default values, as the value in nova.conf will be simply overwrite the default value in nova but not appending the value.

Patch-set: 2