Fix the way Apache is configured for keystone on Ubuntu
On Ubuntu, the site configuration file must have a .conf suffix for
a2ensite and a2dissite to recognise it.

The way it was implemented worked on Fedora, however, because there all
files with the .conf suffix are enabled and such suffix is used by
lib/apache as a mechanism to enable/disable sites.

It fixes the bug by adding a .conf suffix to the file when it is running on
Ubuntu and leaving without suffix on other platforms.

Change-Id: I0328f088f39f1e9a2eb6e73d47ce59059af160fb
Closes-bug: #1313765