Update patch set 4
Patch Set 4:

Ray, when sheepdog is run with the "-c local" option, corosync is not used. Just as corosync is not used if the "-c zookeeper" option is supplied.

I would be happy to add code to build corosync (or zookeeper) from source and make that a configurable option. But I think that would actually reduce the likelihood of this patch merging.

Patch-set: 4
Label: Verified=0