Add switch to configure any capable service under apache
Adds an option ENABLE_ALL_APACHE_SERVICES that will configure all
services that are capable of running under mod_wsgi and apache to
be managed by Apache. The ENABLE_ALL_APACHE_SERVICES needs to be
explicitly set to "1" for this new behavior to occur. This new
option allows for a mechanism to run anything capable under apache
without needing to specifically deliniate each service in the
APACHE_ENABLED_SERVICES option.

Change-Id: Iff9a58030ab3552693d48afa5c9ba801120ec0b3