Update patch set 8
Patch Set 8: Code-Review-1

this definitely needs to be broken up. In addition, the lib/neutron changes are really dense, that setup needs to be refactored into some functions to make more sense of it.

we also really don't want to be pulling an out of tree patch. That goes beyond the scope of devstack.

Patch-set: 8
Label: Code-Review=-1