Update patch set 1
Patch Set 1:

I may want to argue this a bit.  The wording is poor, yes, but so is the alternative.  It's like 'hole flow' theory: what is moving, the electrons or the holes?

You describe the effect of the command, I'm describing what is actually happening.  The reason I prefer describing what is happening (worded better maybe) is that soon after setting up fd3, we change fd1 and fd2 to point at the awk process, and taken in that order the phrasing will lead you to believe that fd3 is now also writing to the awk process.  (yes, that second comment is wrong, too)

How about this:

    # Set fd 3 to a copy of stdout
    exec 3>&1
    if [[ "$VERBOSE" == "True" ]]; then
        # Set fd 1 and 2 to write to the log file
        exec 1> >( awk -v logfile=${LOGFILE} '
                /((set \+o$)|xtrace)/ { next }
# ...
        # Set fd 6 to summary log file
        exec 6> >( tee "${SUMFILE}" )

Patch-set: 1
Label: Verified=0