Update patch set 1
Patch Set 1:

Ok, I think that we probably want a comment in there that this is a work around for a fedora bug which doesn't do IPv6 correctly. As there have been people IPv6 scrubbing devstack recently, and I can a review coming in to delete this in a month and me forgetting why it's here.

Patch-set: 1