Fixed stack.sh environment variables setting issue [bug #1316326]
When a user without proper privilege tries to execute stack.sh to
install Openstack, stack.sh will failed and create a lot of
null environment variables to affect the following scripts
because of no permission to access related files.
This fix will check the exit status to source environments,
if return code not 0, the stack.sh will quit and not produce null
envionment variables to affect next running.
Closes-Bug: 1316326

Change-Id: I4006cb122fa9d850fea8d22cf444f6704709dd2b