Update patch set 1
Patch Set 1:

My plan is to run heat-slow on check and gate for commits to the following projects.

* diskimage-builder
* tripleo-image-elements
* os-collect-config
* os-refresh-config
* os-apply-config
* heat-templates
* heat-cfntools

heat-slow will use the custom image built in the last commit of this series. I haven't done the infra config change to do this yet, but I will get to it. I'd quite like to talk to you in Atlanta about plans and implications for this.

Patch-set: 1
Label: Verified=0