Start nova-conductor before nova-compute
nova-compute always starts with

  "2014-05-09 02:36:34.112 WARNING nova.conductor.api
  [req-98698a08-a3e0-4226-9ec2-f8f22ccf6262 None None] Timed out
  waiting for nova-conductor. Is it running? Or did this service start
  before nova-conductor?"

because it starts before nova-conductor.  Although the connection is
retried, this always leads to a 10-second timeout starting
nova-compute (see the code referenced in [1] for the backoff logic;
the change suggests making this situtation look less like an error).

To avoid this warning and timeout, we should start nova-conductor
first.

[1] https://review.openstack.org/#/c/93375/

Change-Id: I9c877b9acb0cf64aaddc64b8f3f2c27e12419559