Update patch set 1
Patch Set 1: Code-Review-1

I'd like to take a different approach to this...any adjustments that need to be made for either Apache 2.2/2.4 or Ubuntu/Fedora should be made in lib/apache.  The rest of DevStack should have a single way to refer to Apache 'sites' or 'configs'.  The most compatible is with .conf, ie, horizon.conf.  The major change required is to have the fedora handlers not add/remove '.conf' to the site name, but to add/remove and additional suffix like '.disabled' so the names do not have to be parsed.

A quick by-hand test shows that this should work.

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1