Update patch set 3
Patch Set 3:

So the issue here is totally in Grenade in that the changed default in the 'new/target' tree is never downloaded as Grenade copies the images from the 'old/base' tree.  The fix is to either add 0.3.2 to the icehouse devstack list, but not make it the default, so it gets cached; or force the 'new/target' devstack to just use the same image that 'old/base' uses.

Patch-set: 3
Label: Verified=0