Update patch set 2
Patch Set 2: Code-Review-1

actually disagree here. The timeouts are pretty env specific.

For instance, the orchestration image and timeout are coupled, and are set in the gate. The reason this is a timeout is that fedora is sometimes not booting at all in the gate.

We actually really do want these things peeled off, and not tune tempest for the gate environment. We get lots of criticism for that tuning, and they should be decoupled.

Patch-set: 2
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1