Update patch set 2
Abandoned

Given that it seems we have a variance in the size of VMs we use to run tests, i.e. sometimes I see 8 workers in the logs and sometimes 2, I still think the hard-coding is too much of a one size fits all approach, and could be configured a bit smarter depending on the number of CPUs, but I'll go back in my hole for now.

Patch-set: 2
Status: abandoned