Update patch set 1
Patch Set 1:

Architecturally, I'm OK with this -- as long as it's clear that we are
introducing gating on non-integrated projects. FWIW, Ironic is already
effectively gating on a subset of dib's behavior (namely, ramdisk-image-create).

Since this is directly affecting Ironic, I'd like to see the
check-tempest-dsvm-virtual-ironic-nv job pass before we land this.

I'm posting this without a +/- vote to trigger a recheck, before I dig
into why the last run of that test failed.

Patch-set: 1
Label: Verified=0