Bug Fix: 1291111
The installation will be failure if the network proxy is set.
The reason of the failure is that any REST request sent to
the installed OpenStack itself will be redirected to the
proxy server. By default, the no_proxy is set to localhost
only, which needs to be reset to the localhost + HOST_IP +
FLOATING_RANG + FIXED_RANGE.

Change-Id: I6b8e2ba2b7e78fa280283d5ee82d618d4f44ec5a