Reset no_proxy when the node is installed behind a proxy
commit: I3cdace056b74fec111d24d222a886716a063d340
Date: Tue May 27 14:15:00 2014 -0500
Implements: bugfix BUGFIX
Closes-Bug: 1291111/1322200
Reason of the change: The installation will be failure if the network
proxy is set. The reason of the failure is that any REST request sent
to the installed OpenStack itself will be redirected to the proxy
server. By default, the no_proxy is set to localhost only, which needs
to be reset to the localhost + HOST_IP + FLOATING_RANG + FIXED_RANGE.
Changes:
    Add "export_no_proxy_variables" in stack.sh
    This is used to reset the no_proxy variable after HOST_IP,
FLOATING_RANG, FIXED_RANGE have been set.

    Add a new function "export_no_proxy_variables" in function_common
The new function will reset the no_proxy variable. And save it in the
shell configuration file ".bashrc" and "localrc".

Change-Id: Ifa4ec3f4b0d42a615ea18a1cdb4e24854e5f72ca