Reset no_proxy when the node is installed behind a proxy
Date: Fri Nov 07 16:36:00 2014 -0500
Implements: bugfix BUGFIX

Closes-Bug: 1322200

Problem:
=======
When devstack is called by Trove for installation behind a proxy,
the no proxy variables are set by stack.sh. but it doesn't passed
back to Trove. Thereforethe the installation will fail.

Solution:
========
Save the no_proxy variables in localrc.
This fix is based on the fix for 1291111

Change-Id: Ie7058600266f033af0836bdc4de5a56e823f8c07