Reset no_proxy when the node is installed behind a proxy
commit:
Author: Zu Qiang
Date: Thu May 22 11:35:00 2014 -0500
Implements: bugfix BUGFIX
Closes-Bug: 1291111/1322200
Reason of the change: The installation will be failure if the network
proxy is set. The reason of the failure is that any REST request sent
to the installed OpenStack itself will be redirected to the proxy
server. By default, the no_proxy is set to localhost only, which needs
to be reset to the localhost + HOST_IP + FLOATING_RANG + FIXED_RANGE.
Changes:
 Add "call export_no_proxy_variables" in stack.sh
     This is used to reset the no_proxy variable after HOST_IP,
FLOATING_RANG, FIXED_RANGE have been set.
     # Reset no_porxy variables
     export_no_proxy_variables
 Add a new function "export_no_proxy_variables" in function_common
The new function will reset the no_proxy variable. And save it in the
shell configuration file. It also creates a new file "noproxy".
The noproxy file will be used by other installation procedure,
e.g. Trove.
   # Update no proxy environment variables
   # no_proxy is set in $USERHOME/.bashrc and $USERHOME/noproxy
   #
   function export_no_proxy_variables {
       export no_proxy=`echo $no_proxy,$HOST_IP,$FLOATING_RANG,$FIXED_RANGE | sed -e 's/,\+/,/g'`
     # Check if the USERHOME is set
       if [[ -z "$USERHOME" ]]; then
           USERHOME=$HOME
       fi
     # Save the no_proxy configuration in the shell rc
       if [[ -n `grep 'no_proxy' $USERHOME/.bashrc` ]]; then
          sed -i "/no_proxy=/c\no_proxy=$no_proxy" $USERHOME/.bashrc
       else
          echo -e "# Reset the no_proxy variables\nno_proxy="$no_proxy >> $USERHOME/.bashrc
       fi
     # Recreating no_proxy file
       echo no_proxy=$no_proxy > $USERHOME/noproxy
       chmod +x $USERHOME/noproxy
   }

Change-Id: I3cdace056b74fec111d24d222a886716a063d340