Reset no_proxy when the node is installed behind a proxy
Date: Mon Jun 04 13:25:00 2014 -0500
Implements: bugfix
Closes-Bug: 1291111/1322200

Reason of the change:
When the Openstack is installed with a network proxy,
the installation will be failure. The reason is that any REST request
sent to Keystone will be redirected to the network proxy, which will be
dropped as normally the local host address is unknown if the Openstack
is installed behind a NAT.
Here is the error code at installation: "ERROR: cliff.app Service
Unavailable (HTTP 503)"
By default, the no_proxy is set to localhost only. The OpenStack
address is not added in the no_proxy parameter.

The Changes are:

 Add call export_no_proxy_variables in stack.sh. This is used to
reset the no_proxy variable after HOST_IP, FLOATING_RANG,
FIXED_RANGE have been set.

 Add a new function "export_no_proxy_variables" in function_common
The new function will reset the no_proxy variable to localhost +
HOST_IP + FLOATING_RANG + FIXED_RANGE. And save it in the shell
configuration file and localrc.

Change-Id: I3cdace056b74fec111d24d222a886716a063d340