Update patch set 1
Patch Set 1: Code-Review-1

This is still failing the virtual-ironic job -- which is mostly stable at this point.

From looking in the logs, it appears the problem is that nova-compute was started 35 seconds before ironic-api, leading to n-cpu failing during init.

This is recorded at the following bug but hasn't been a problem for this job so far. It showed up in other testing.

https://bugs.launchpad.net/ironic/+bug/1309548


I'm not sure why this change would change the order that n-cpu and ir-api are started in, but it seems to have done just that.

Patch-set: 1
Label: Code-Review=-1