Update patch set 6
Patch Set 6:

I do not believe so. I added the lines to ensure the interfaces are turned on - it could be that it was just because they were not configured properly in /etc/network/interfaces - I don't recall the exact circumstances around it - but it happened enough that I wanted to always make sure they were turned on properly.

Patch-set: 6