Update patch set 1
Patch Set 1:

What I'm trying to do is adjust your devstack-vagrant setup to work with fedora20 (and eventually mongodb for ceilometer) as a starting point and during one of several attempts at that I saw the chown command fail with an empty $USER. I thought, well that's plenty weird and cast about for a way to make sure that couldn't happen. Looking elsewhere in the code it was clear that $STACK_USER was the preferred way to go.

Patch-set: 1
Reviewer: Gerrit User 11564 <11564@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0