soft-qemu handling with F20 could image
When the qume used with Linux image without
a para-virtualized timer, various timing issues can happen,
if the qemu process preempted in the wrong time, for example
at timer test or at loop per jiffy calculation.

The issues less like happens on low load, high core number
host system, but it can happen.

For soft qemu (TCG) generally recommended to explicitly disable,
the tsc and timer_check, it could be done at image building time, but
the loops per jiffy value varies on different cpu,
and cannot be done at build time.

It is not the nicest patch ever, but it is required.

Related-Bug: #1297560
Closes-Bug: #1312199

Change-Id: Id5cd01a92a047b7859914e5bb017c15ee443b4d5