Update patch set 5
Patch Set 5:

IIRC, the origin author had this in a list/dict because the extension name, and the migration name were different for some extensions. But this has been fixed in Keystone, it's now predictably <name> and <name>_extension.

I think it's best used as just a comma separated string, and not a list, (like ENABLED_SERVICES).

Patch-set: 5