Update patch set 7
Patch Set 7: Code-Review-1

(2 comments)

I'm kind of concerned that extensions require modifying the paste ini, which lives in this weird space between config and code, which is also going to make for interestingness in upgrade. I'm also surprised that the defaults don't include any?

Is there some documentation about how this all is expected to work on the keystone side? Just because this is so different from the Nova or Neutron extensions that I'm having a hard time evaluating if this is how it would make sense to integrate into devstack with sensible defaults.

Patch-set: 7
Label: Code-Review=-1