Update patch set 4
Patch Set 4:

what's the minimum version of the middleware that is required for this?

Also any chance that we can clean up the commenting out of the paste files? Like have Nova and Cinder started removing those parts? I don't like us having to delete so much of the default config for this to work.

Patch-set: 4
Label: Verified=0