Update patch set 5
Patch Set 5:

Minimum version added.

I removed the comment changes to nova and cinder, the identity_uri is not present in either of those cases i was just aiming for consistency. 

In the nova case it would appear that none of the commenting is required. Cinder still carries auth_host etc options.

Patch-set: 5
Label: Verified=0