Creates randomized fixed_key in stack.sh
The key manager in nova is currently configured to use a fixed key for
testing until a proper key manager is integrated.  However, encrypted
volumes will fail during attachment to a server if a fixed_key is not defined.
Randomized creation of a fixed_key in stack.sh was included in the
stable/icehouse release for testing purposes.

Adding a randomized value for fixed_key allowed new encryption tempest test CR
to pass in the gate for tests that ran the stable/icehouse branch of Devstack.
There is still an issue with tests that start off running stable/havana, such
as the Icehouse Grenade tests.  See: https://review.openstack.org/#/c/58018/

If the randomized value for fixed_key is also added to the stable/havana
branch, the patch linked above should pass all of the Jenkins tests.

Change-Id: Ieae6ed056c5e51cda7509a87cbe9b02586c6365b