Update patch set 5
Patch Set 5:

I tried to do sahara and had a number of problems. I was going to attempt it as another review. 

I *think* it is a problem of bug 1323934 - if you know sahara better and want to investigate further have a look at the jenkins reports around patch 2 & 3. 

The comment on the bug is that they are trying to update to the new config options internally and i think we will have to wait for that, see bug 1257472 for that work.

Patch-set: 5