Add IPv6 management network support
This changeset adds support of IPv6 management network in devstack. Define
IP_MGMT_NET={ipv6 | ipv4} in your localrc to indicate if IPv6 or IPv4
management network shoule be used. Default management network is ipv4.

Note that there is a bug in eventlet that prevents the use of literal IPv6
addresses in URL. The fix to it hasn't been released. Therefore, it's
recommended that HOST_IP and SERVICE_HOST be explicitly defined with HOSTNAMES.

when literal IPv6 addresses are supported, make sure that HOST_IP and
SERVICE_HOST are using literal IPv6 address format when defining them in your
localrc. That is, quote the IPv6 address with '[' and ']'.

Also in your localrc, make sure to disable n-api-meta. There is a bug there
that prevents nova api from being started successfully.

Change-Id: Id913f8e25d75813de1c56abb63102b86f6dec0b5
Implements: blueprint ipv6-support