Use tempest default for ssh_timeout value
We've seen a spike in ssh timeout failures since 6/5/14, possibly
related to Tempest commit b7f648b31c9044e3a7c25c0fb1dcc31a81e0da9f which
adds more servers to the test runs.  The default ssh_timeout in Tempest
right now is 300 seconds and this removes setting it to build_timeout
which is configured to be 196 seconds in devstack.

Related-Bug: #1298472

Change-Id: Ide0cb68898c944f852b8bbbda8bf564038c3d351