Set osapi_volume_workers equal to number of CPUs
Cinder blueprint multi-process-api-service provided a way to configure
the number of Cinder API workers (similar to Nova). By default it sets 1
worker like Glance and Ceilometer.

Nova changed it's default API workers equal to the number of CPUs by
default, so it's logical to do the same here for Cinder.

There is also hope that this might help with some latent volume delete
leaks in Tempest.

Related-Bug: #1322011

Change-Id: Iaa28d98b9492044f6c227b78ae0e22451cbc4f74