Update patch set 4
Patch Set 4:

There's a lot of random failure here.  The one thing in common seems to be that devstack exits with a 127 error code, but it's not leaving the real problem in the logs.  It always seems to be *around* the time of your change here, which makes me wonder if the logs haven't been flushed when it exited.

127 suggests a command not found; the only command you added was "bc" so I wonder if that's not always around?

What if you try doing it like [1] which doesn't rely on this?

[1] https://github.com/openstack-dev/devstack/blob/master/stack.sh#L412

Patch-set: 4
Label: Verified=0