Expand variables in get_meta_section_files
get_meta_section_files has an implied API of returning each
config-file that is affected by a meta-section command on a new-line.
Thus it should be expanding any variables it might find before
returning them, for example

 [[post-config|/$Q_PLUGIN_CONF_FILE]]

To facilitate this, run eval() over the filename to expand it fully.
Unfortunately, eval() seems like the only option without starting to
replicate the semantics of eval itself.  Nothing implies that the
file-name should start $ if it is a variable (see counter example
above), or N variables ($TOP_DIR/$CONFIG/$FILE...) isn't valid, etc.

A test-case is added to test_config.sh

Closes-bug: #1294213
Change-Id: Id0b441043f8eb5370107e419cdd183b4baa4c863