Wait for OVS to reconfigure when creating PUBLIC_BRIDGE
The '--no-wait' flag will not wait for ovs-vswitchd to reconfigure after
changes have been made to the database. This causes a race condition as
the bridge may or may not be up when 'ip addr flush dev $PUBLIC_BRIDGE' is
called. This commit removes the '--no-wait' flag which is safe as the
openvswitch services are already started earlier in stack.sh

Change-Id: I73ef1b731295d03b0f064c5f786196284c3e39ce
Closes-bug: #1328962