Update patch set 3
Patch Set 3:

You could set the correct label on the horizon files and directories (http://fedoraproject.org/wiki/SELinux/apache), but
horizon also requires to the httpd_t to be able to connect to commplex_main_port_t (5000) keystone.
osapi_compute_port_t (8774) nova.
neutron_port_t ( 9696, 9697).

For having the swift's rsync working probably rsync_full_access sebool needs to be turned on, or the security labels need to be set on the fs used by swift. (The test cases does not sees if the rsync not working)

For now I would recommend to set the selinux mode to permissive.
In this case we will not have failures related to avc denial and
 the selinux policy violation still be logged in /var/log/audit/audit.log*

In tools/fixup_stuff.sh the
    # Disable selinux to avoid configuring to allow Apache access
    # to Horizon files (LP#1175444)
    if selinuxenabled; then
        sudo setenforce 0
    fi

should be extended to all fedoras.

Patch-set: 3
Label: Verified=0