Set glance-api workers equal to number of CPUs
Glance allows configuring the number of API workers and recommends
setting the value equal to the number of CPUs available on the host, but
defaults to 1 in the code.

This change sets glance-api workers equal to the number of CPUs based on
the recommendation in the config in the hopes of improving performance
of Tempest testing.

Change-Id: I53039e4032b398a6d1fef9b8a5485d7c1204c749