Reserve port 35357 from the ephemeral range
Reserve port 35357 from the ephemeral range as early as reasonably
possible in the stack.sh process to reduce the likelyhood that the
port will be in use. This does not completely resolve the issue
where Keystone's IANA assigned port falls into Linux's ephemeral
range, but this should reduce the occurances.

Change-Id: I8cfb53d8f90c1ff1fb1083c59fefabca3d14323b
Partial-Bug: #1253482