Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Thanks for doing this after we've talked about it for so long...

Since stack.sh often runs multiple times per system boot (on non-CI systems anyway) this should probably also check to see if 35357 is already reserved and do nothing if so.  Duplicates are probably mostly harmless but it should be clean.

Also, you can call reserve_keystone_port() from init_keystone() and not need to add it to stack.sh as it is a Keystone-only thing.

Patch-set: 2
Label: Code-Review=-1