Update patch set 2
Patch Set 2:

I am adding it as a keystone only thing in stack.sh so we do it before we have extra stuff "running" on the system.  if this _really_ needs to go in keystone init, i think we're not eliminating the problem, because by the time we hit keystone init something could be using that port. The earlier in the process we reserve the port, the less likely it will be in use when keystone tries and starts.

Unfortunately when it comes to detecting if it's already there the kernel likes to emit "ranges".

You could have ports 30000-39999,42675 as the return value of the reserved ports (for example). If you re-reserve it, the kernel will just smush down the multiple reservations into a range (if possible) or ignore it. It doesn't hurt to reserve it over and over.

Patch-set: 2
Reviewer: Gerrit User 2903 <2903@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0