Update patch set 2
Patch Set 2:

I'll get it in the fixup scripts, this really is more along those lines (I dislike it being here at all, but what can we do?). I don't think it's OpenStack stuff starting that is causing the issue, i think it's something else running that is consuming an ephemeral port that just happens to be keystone's. 

I'll get the variables used for the port instead of the number(s) in the next patch.

Thanks for the comments!

Patch-set: 2