Use "_member_" role instead of "Member"
As Horizon now defaults to using the _member_ role, automatically
created by Keystone, there is no need to create a Member role.

This commit disables the creation of this additional role and modifies
the remaining dependencies on the Member role within devstack to use
_member_.

Additionally as all users who are created with an association to a
tenant are automatically given the _member_ role in that tenant, manual
inititiation of the role add command to explicitly add the role is
also no longer required.

Conflicts:
	lib/keystone
	lib/tempest

Change-Id: I29af9fb5dc52bc0af57509c1c0fed2345231f649
Closes-Bug: #1330132