Update patch set 3
Patch Set 3:

> Does get_meta_section() work fine?

This is a very good point :)  The only way I can see for awk to know that is to use an exported variable which is no good.

In [3] revision 3 I have expanded the test cases to cover get_meta_section and merge_config_group.  I presume this passes with them?  I think this is complex and fragile enough that we want full test-cases before we merge anything.


[3] https://review.openstack.org/#/c/98951/

Patch-set: 3