Update patch set 3
Patch Set 3: Code-Review+2

I'm reluctant to drop the memory on the heat flavor, but if we think it's going to help, so be it.

Have we considered using the AMI for fedora instead of the disk image? To avoid the emulated bios setup?

Patch-set: 3
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2