Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

realistically, I don't want that level of granularity if we are doing SSL.

I want 'USE_SSL=True' (and I eventually want it to be default true)

And that's the end of it. And I'd like us to remain with a single CA / key for all the services for now, given that our default is an all in one setup, unless there is a really big reason not to.

Patch-set: 2
Label: Code-Review=-1