Update patch set 1
Patch Set 1: Code-Review-1

Can we do this a little differently. Use the value of /proc/sys/net/ipv6/conf/default/disable_ipv6

To figure out if we should use ipv6 on the host. If it's not set, use it.

Also, from my reading :: as a bind is all of ipv4 and ipv6, so I think we can just set that unconditionally.

I actually want ipv6 things defaulting on if we can manage them, defaulting off is going to mean less testing on them.

Patch-set: 1
Label: Code-Review=-1