Update patch set 2
Patch Set 2:

USE_SSL could be a very misleading option.

Native SSL is not considered a viable deploymen option by many. Given that the code to do so is available and running it alleviates the need for a slew of stud processes I went ahead and implemented it. That said, the OSSG promotes the use of TLS proxy over native SSL so we may want to encourage that as a use case over native.

Enabling SSL via the proxy is as easy as adding a single server to the enabled list.

Enabling native SSL is purposefully more difficult. I can certainly change it to use the same cert/key/CA file everywhere but I was keeping with the model that Jamie set out. I'd want to rip that out completely if we go with a single-cert solution. It would also be fairly trivial to execute the same CA-creation calls that the TLS Proxy code already uses.

I'd like some feedback before I start ripping out code.

Patch-set: 2
Label: Verified=0