Update patch set 2
Patch Set 2:

I'd certainly prefer to use the existing CA bits everywhere.  If there is a shortcoming in that let's fix it rather than do something else.

The continued value of native SSL in the servers is not obvious to me.  They are not consistent in support or configuration, nor do I believe an actual production deployment of OpenStack will use it.  I'd like to know there is actual value in spending testing resources on that before committing to it.

I think we're better off using a proxy or running WSGI under Apache to get SSL, and honestly, if we could do that with Apache the value of the stud proxy implementation goes down...

Patch-set: 2
Label: Workflow=0