Update patch set 2
Patch Set 2:

The native SSL configuration code isn't a whole lot different than the TLS Proxy code so IMHO it won't be a testing burden. There are actually uses doing native SSL in production now (for good or bad).

The OSSG promotes the use of TLS proxies for more than just avoiding python SSL (load balancing and such). I'm not set on stud as the proxy, just tried to keep the changes as limited as possible. The tls-proxy part is abstracted well enough that changing the actual proxy should be a no-brainer.

Patch-set: 2