Update patch set 5
Patch Set 5: Code-Review-1

I noticed this failed with Fedora.

The -L is required, but it turns out there is an issue [1] that may or may not be related to apache/github but can be worked around using --no-sessionid; otherwise this returns a 400 error.

I know this isn't your main concern, but I think since it's identified we should deal with it.  I think adding --no-sessionid is safe on all platforms and it should be noted with a reference to the bug

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1098711

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1