Update patch set 7
Patch Set 7:

The fact that the nodes can't reach the TFTP server appears to be a bug with Devstack's deployment of Neutron.  

I was under the assumption that the neutron linux-bridge agent was fully supported in Devstack.  

The create_neutron_initial_network function in lib/neutron is configuring the gateway IP address on the $PUBLIC_BRIDGE interface (br-ex) only when the OVS plugin is being used and the q_l3 service is enabled.  Is it wrong to assume that devstack should also be setting up the gateway IP address in a similar fashion when the linuxbridge agent is being used?

Patch-set: 7
Reviewer: Gerrit User 7881 <7881@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0