Update patch set 22
Patch Set 22: Code-Review-1

(3 comments)

sorry for the delay here. Finally got around. There are a few cleanups mentioned in line that I think probably need to be addressed.

One other question: the create_disk function, can the swift paths be plumbed on top of that as well? I'd rather not have 2 functions creating xfs loopbacks differently.

Lastly, the one thing that still doesn't feel right here to me is the short circuits in lib/cinder. It seems like we should actually extract the tgt/lvm path into a plugin (which is the default backend), and then that would go away.

As I wasn't in the room with the discussion with dean, I'm not sure what else was agreed to. So I'm happy to be trumped by his POV.

Patch-set: 22
Label: Code-Review=-1