Update patch set 7
Patch Set 7:

Anthony, I am not super familiar with how devstack does Neutron, but in the context of baremetal I believe that is correct. We need a gateway to be set via DHCP for nodes netbooting within the tenant namespace, so traffic is routeable to the TFTP server to the 'external' management network (@ $HOST_IP)  There looks to be some special casing for baremetal in create_initial_neutron_network() that could be updated for the linuxbridge case?   If it requires more general fixing, perhaps you can create another patch and make this review dependent on that one.

Patch-set: 7