Update patch set 6
Patch Set 6:

In the discussion on IRC with Kyle I argued as follows:

Note that ovs-base is not an explicitly enabled service.
It depends on the q-agt service, and not all q-agt need OVS.
OpenDayLight needs OVS on compute nodes.
Neutron may need OVS on compute and/or server nodes.

With this patch, when we get to the point where we know we need to install OVS, we do so.
We are installing OVS, so the apts/rpms are for OVS (not the g-agt or ODL).

In the future if ODL needs some other apts/rpms not related to OVS, it can add a file/apts/odl-compute for them and they will get installed automatically.

Patch-set: 6