Update patch set 16
Patch Set 16: Code-Review-1

(6 comments)

OpenStackClient 0.4.0 was released last week that forces cliff to a new enough version to make the -f value option work properly.  This lets the places where a single value is desired, like a user ID, get it directly and not require the grep | get_field pipeline.

All of the get_or_create_*() functions can be simplified to use the form shown in get_or_create_user().  This really needs to be done as much for demonstrating the intended use case as for simplifying DevStack.

for example, in the case of Keystone backed by a corporate LDAP, 'user list' will most likely return a small subset of the complete user list.  User existence discovery must be done with 'user show'.

Patch-set: 16
Label: Code-Review=-1